Make Fake test classes reusable
