+++ src/test/java/com/netflix/playback/features/model/FakeKeyLoggerHandler.java
@@ -1,0 +1,14 @@
+package com.netflix.playback.features.model;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+public class FakeKeyLoggerHandler implements KeyLogger.Handler {
+  public final List<String> allKeys = new ArrayList<String>();
+  
+  @Override
+  public void handle(String... keys) {
+    allKeys.addAll(Arrays.asList(keys));
+  }
+}
