+++ src/test/java/com/netflix/playback/features/model/KeyLoggerTest.java
@@ -2,24 +2,11 @@ package com.netflix.playback.features.model;
 
 import static org.junit.Assert.assertArrayEquals;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 import org.junit.Before;
 import org.junit.Test;
 
 public class KeyLoggerTest {
   private KeyLogger keyLogger = new KeyLogger();
-  
-  private class FakeHandler implements KeyLogger.Handler {
-    private final List<String> allKeys = new ArrayList<String>();
-    
-    @Override
-    public void handle(String... keys) {
-      allKeys.addAll(Arrays.asList(keys));
-    }
-  }
   
   @Before
   public void setUp() throws Exception {
@@ -33,8 +20,8 @@ public class KeyLoggerTest {
 
   @Test
   public void testMultpleHandlers() {
-    FakeHandler handlerA = new FakeHandler();
-    FakeHandler handlerB = new FakeHandler();
+    FakeKeyLoggerHandler handlerA = new FakeKeyLoggerHandler();
+    FakeKeyLoggerHandler handlerB = new FakeKeyLoggerHandler();
     keyLogger.addHandler(handlerA);
     keyLogger.addHandler(handlerB);
     keyLogger.log("a", "b", "c");
